feat: Added a player choice validation system
nested the code in playRound() inside a while loop that has condition; !validOption. That means it only works if validOption is false. And then added continue statemnt if player chooses an invalid option. So, it keeps on repeating the code and skipping the code that assigns validOption true which was the only way to end the playRound function. So, this keeps repeating until the player chooses a valid option and the code doesn't encounter a continue statement.