refactor: Change code to get playerSelection from Buttons
Instead of getting playerSelection from prompt and then converting it to be case insensitive, now playerSelection is assigned depending on the id of the target event(button click). The console logs the result of the game depending on the selections of both players and the playRound() function is now triggered after the button gets clicked.